requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.11.0
django-crontab==0.7.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.0.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
